최종배포

+++ src/main/java/com/finalproject/breeding/image/AwsS3Service.java
@@ -14,9 +14,9 @@ import com.finalproject.breeding.image.repository.TogetherImageRepository;
 import com.finalproject.breeding.image.repository.UserImageRepository;
 import io.jsonwebtoken.io.IOException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.util.ObjectUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
@@ -27,6 +27,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@Slf4j
 @RequiredArgsConstructor
 public class AwsS3Service {
 
@@ -39,7 +40,16 @@ public class AwsS3Service {
     @Value("${cloud.aws.s3.bucket}")
     private String bucket;
 
-//    public PostImage uploadPost(MultipartFile multipartFile, String dirName) throws IOException, java.io.IOException {
+    /*ArrayList<String> accessableLists = new ArrayList<>();
+    public ArrayList<String> setAccessableLists() {
+        accessableLists.add(".jpg");
+        accessableLists.add(".png");
+        accessableLists.add(".gif");
+
+        return accessableLists;
+    }*/
+
+    //    public PostImage uploadPost(MultipartFile multipartFile, String dirName) throws IOException, java.io.IOException {
 //        File file = convertMultipartFileToFile(multipartFile)
 //                .orElseThrow(() -> new IllegalArgumentException("MultipartFile -> File convert fail"));
 //
@@ -55,9 +65,25 @@ public class AwsS3Service {
 //                .path(path)
 //                .build();
 //    }
+
     public List<PostImage> uploadPost(List<MultipartFile> multipartFiles, String dirName) throws IOException, java.io.IOException{
 
+        // 파일 확장자명 벨리데이션 체크
+        ArrayList<String> accessableLists = new ArrayList<>();
+        accessableLists.add(".jpg");
+        accessableLists.add(".jpeg");
+        accessableLists.add(".png");
+        accessableLists.add(".gif");
+
         List<PostImage> postImages = new ArrayList<>();
+
+        for (MultipartFile multipartFile : multipartFiles) {
+            String ext = multipartFile.getOriginalFilename().substring(multipartFile.getOriginalFilename().lastIndexOf("."));
+            if (!accessableLists.contains(ext)) {
+                log.error("500 Error : 사진 파일이 아니거나, 지원하지 않는 확장 파일입니다.");
+                throw new CustomException(ErrorCode.EXTRACTION_VALIDATION_ERROR);
+            }
+        }
 
        for (MultipartFile multipartFile : multipartFiles){
            File file = convertMultipartFileToFile(multipartFile).orElseThrow(()->new CustomException(ErrorCode.IMAGE_UPLOAD_ERROR));
@@ -99,9 +125,21 @@ public class AwsS3Service {
     }
 
     public UserImage uploadUser(MultipartFile multipartFile, String dirName) throws IOException, java.io.IOException {
+        ArrayList<String> accessableLists = new ArrayList<>();
+        accessableLists.add(".jpg");
+        accessableLists.add(".jpeg");
+        accessableLists.add(".png");
+        accessableLists.add(".gif");
+
         File file = convertMultipartFileToFile(multipartFile)
                 .orElseThrow(() -> new CustomException(ErrorCode.IMAGE_UPLOAD_ERROR));
 
+        String ext = file.getName().substring(file.getName().lastIndexOf("."));
+
+        if(!accessableLists.contains(ext)){
+            log.error("500 Error : 사진 파일이 아니거나, 지원하지 않는 확장 파일입니다.");
+            throw new CustomException(ErrorCode.EXTRACTION_VALIDATION_ERROR);
+        }
         String key = randomFileName(file, dirName);
         String path = putS3(file, key);
         removeFile(file);
@@ -159,6 +197,13 @@ public class AwsS3Service {
         for(CommunityImage communityImage : communityImages){
             remove(communityImage.getKey());
             communityImageRepository.delete(communityImage);
+        }
+    }
+    public void removeTogetherImages(Long togetherId) {
+        List<TogetherImage> togetherImages = togetherImageRepository.findByTogetherId(togetherId);
+        for(TogetherImage togetherImage : togetherImages){
+            remove(togetherImage.getKey());
+            togetherImageRepository.delete(togetherImage);
         }
     }
     public void removeUserImage(Long userId){
